refactor: 전반적인 코드 수정
