+++ src/main/java/cart/Repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package cart.Repository;
 
+import cart.Repository.mapper.ProductMapper;
 import cart.dao.ProductDao;
 import cart.domain.Product.Product;
 import cart.entity.ProductEntity;
