+++ src/main/java/cart/WebMvcConfig.java
@@ -1,7 +1,7 @@
 package cart;
 
 import cart.Repository.MemberRepository;
-import cart.ui.MemberArgumentResolver;
+import cart.auth.MemberArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
