+++ src/main/java/cart/domain/Member/Email.java
@@ -11,6 +11,7 @@ public class Email {
     private final String email;
 
     public Email(final String email) {
+
         this.email = email;
     }
